+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added global WMO minimap preview to terrainless maps in the map viewer.
 - Added "Export WMO Minimap" option to export global WMO minimap as PNG.
 - Added option to export individual character textures from the character viewer.
+- Added option to export models and characters (posed) in STL format (for 3D printing).
 - Cinematic playback in videos tab now includes audio and subtitles.
 - Optimized "Maps" tab to load 33x (97%) faster.
 - Optimized map viewer to load initial tiles must faster.